Extracted NavigationBar to separate component
