+++ src/App.js
@@ -1,25 +1,12 @@
 import { HashRouter, Switch, Route, Redirect } from "react-router-dom";
+import NavigationBar from "./core/NavigationBar";
 import TaskPage from "./features/tasks/TaskPage";
 import TasksPage from "./features/tasks/TasksPage/index"
 import AuthorPage from "./features/author/AuthorPage/index";
-import { NavigationBar, NavigationItem, NavigationWrapper, StyledNavigationLink } from "./styled";
 
 const App = () => (
     <HashRouter>
-        <NavigationBar>
-            <NavigationWrapper>
-                <NavigationItem>
-                    <StyledNavigationLink to="/zadania">
-                        Zadania
-                    </StyledNavigationLink>
-                </NavigationItem>
-                <NavigationItem>
-                    <StyledNavigationLink to="/author">
-                        O autorze
-                    </StyledNavigationLink>
-                </NavigationItem>
-            </NavigationWrapper>
-        </NavigationBar>
+        <NavigationBar />
 
         <Switch>
             <Route path="/zadania/:id">
